requirements.txt
phidata 
python-dotenv
packaging
fastapi
uvicorn
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
generative-ai-hub-sdk
ai-core-sdk